% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GDALSQL.R
\docType{methods}
\name{dbFetch,GDALSQLResult-method}
\alias{dbFetch,GDALSQLResult-method}
\title{Retrieve records from GDALSQL query}
\usage{
\S4method{dbFetch}{GDALSQLResult}(res, n = -1, ...)
}
\description{
Retrieve records from GDALSQL query
}
