% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GDALSQL.R
\docType{methods}
\name{dbSendQuery,GDALSQLConnection-method}
\alias{dbSendQuery,GDALSQLConnection-method}
\title{Send a query to GDALSQL.}
\usage{
\S4method{dbSendQuery}{GDALSQLConnection}(conn, statement, ...)
}
\arguments{
\item{conn}{database connection, s created by \code{\link{dbConnect}}}

\item{statement}{OGR SQL, see http://www.gdal.org/ogr_sql.html}

\item{...}{for compatibility with generic}
}
\description{
Send a query to GDALSQL.
}
\examples{
#f = system.file("example-data/continents", package = "rgdal2")
#chuck <- rgdal::ogrInfo(f, "continent") ## bizarrely this resets the problem
afile <- system.file("extdata", "shapes.gpkg", package = "RGDALSQL")
db <- dbConnect(RGDALSQL::GDALSQL(), afile)
dbSendQuery(db, "SELECT * FROM sids WHERE FID < 1")
}
