% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GDALSQL.R
\docType{class}
\name{GDALSQLDriver-class}
\alias{GDALSQLDriver-class}
\alias{dbUnloadDriver,GDALSQLDriver-method}
\title{Driver for GDAL SQL.}
\usage{
\S4method{dbUnloadDriver}{GDALSQLDriver}(drv, ...)
}
\description{
Driver for GDAL SQL.
}
\keyword{internal}
