% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfr.R
\name{read_rgdal}
\alias{read_rgdal}
\title{Read raw rgdal2 output.}
\usage{
read_rgdal(x, layer = 1L)
}
\arguments{
\item{x}{data source name (e.g. a file, or a database connection string)}

\item{layer}{a layer name from the data source, or a 1-based index of those}
}
\description{
Read a list of the geometry and the object metadata as nested lists.
Includes a `meta` table to store the CRS string/s, the source name/s, plus future stuff.
}
\details{
sf processes this all in-place into R level simple feature classes.
}
