% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfr.R
\name{read_tables}
\alias{read_tables}
\title{simplification of sf::read.sf to avoid classes
need to return the data_data still}
\usage{
read_tables(x, layer = 1L)
}
\description{
simplification of sf::read.sf to avoid classes
need to return the data_data still
}

