% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfr.R
\name{readObject}
\alias{readObject}
\title{Read a single object.}
\usage{
readObject(layer, id)
}
\arguments{
\item{layer}{OGRLayer from rgdal2}

\item{id}{zero based object ID}
}
\description{
Read from a vector layer, returns a single object by (counting) ID.
}
\details{
Flattened to nested list of tibbles. sf calls this "readFeature", and returns matrices.
}

