% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GDALSQL.R
\docType{methods}
\name{dbSendQuery,GDALSQLConnection-method}
\alias{dbSendQuery,GDALSQLConnection-method}
\title{Send a query to GDALSQL.}
\usage{
\S4method{dbSendQuery}{GDALSQLConnection}(conn, statement, ...)
}
\description{
Send a query to GDALSQL.
}
\examples{
f = system.file("example-data/continents", package = "rgdal2")
db <- dbConnect(RGDALSQL::GDALSQL(), f)
dbSendQuery(db, "SELECT * FROM continent WHERE FID < 1")
}

