% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GDALSQL.R
\docType{methods}
\name{dbConnect,GDALSQLDriver-method}
\alias{dbConnect,GDALSQLDriver-method}
\title{dbConnect}
\usage{
\S4method{dbConnect}{GDALSQLDriver}(drv, dbname = "", readonly = TRUE, ...)
}
\arguments{
\item{drv}{GDALSQLDriver}

\item{dbname}{data source name, file, or folder path}

\item{readonly}{open in readonly mode?}
}
\description{
dbConnect
}
\examples{
\dontrun{
db <- dbConnect(RGDALSQL::GDALSQL(), afile)
afile <- system.file("extdata", "shapes.gpkg", package = "RGDALSQL")
chuck <- rgdal::readOGR(afile, "sids") ## bizarrely this resets the problem
con <- rgdal2::openOGR(afile)
#dbWriteTable(db, "mtcars", mtcars)
#dbGetQuery(db, "SELECT * FROM mtcars WHERE cyl == 4")
}
}

