% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GDALSQL.R
\name{GDALSQL}
\alias{GDALSQL}
\alias{dbConnect,GDALSQLDriver-method}
\title{GDALSQL}
\usage{
GDALSQL()

\S4method{dbConnect}{GDALSQLDriver}(drv, DSN = "", readonly = TRUE, ...)
}
\arguments{
\item{drv}{GDALSQLDriver}

\item{DSN}{data source name, may be a file, or folder path, database connection string, or URL}

\item{readonly}{open in readonly mode?}
}
\description{
GDALSQL driver

dbConnect
}
\examples{
\dontrun{
## this is a nothing connection
db <- dbConnect(RGDALSQL::GDALSQL())
afile <- system.file("extdata", "shapes.gpkg", package = "RGDALSQL")
db <- dbConnect(RGDALSQL::GDALSQL(), afile)
dbSendQuery(db, "SELECT * FROM sids")
}
}
