% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GDALSQL.R
\docType{class}
\name{GDALSQLConnection-class}
\alias{GDALSQLConnection-class}
\alias{show,GDALSQLConnection-method}
\title{GDALSQL connection class.}
\usage{
\S4method{show}{GDALSQLConnection}(object)
}
\description{
GDALSQL connection class.
}
\keyword{internal}

